updated field userID

+++ src/main/java/com/spring/ecommerce/lvtn/service/IMPL/UserSericeIMPL.java
@@ -5,6 +5,7 @@ import com.spring.ecommerce.lvtn.model.Entity.User;
 import com.spring.ecommerce.lvtn.repository.UserRepository;
 import com.spring.ecommerce.lvtn.service.UploadFileService;
 import com.spring.ecommerce.lvtn.service.UserService;
+import com.spring.ecommerce.lvtn.service.Util.SequenceGeneratorService;
 import com.spring.ecommerce.lvtn.utils.enums.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,8 +19,11 @@ public class UserSericeIMPL implements UserService {
     private final UserRepository userRepository;
     private final UploadFileService uploadFileService;
     @Override
-    public User uploadAvatar(String id, MultipartFile avatar) throws IOException {
-        User user = userRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
+    public User uploadAvatar(Long userId, MultipartFile avatar) throws IOException {
+        User user = userRepository.findByUserId(userId);
+        if (user == null){
+            throw new AppException(ErrorCode.USER_NOT_FOUND);
+        }
         String oldAvatar = user.getAvatar();
         if (oldAvatar != null){
             uploadFileService.deleteFileByUrl(oldAvatar);
